+++ src/main/java/com/google/step/snippet/external/StackOverflowClient.java
@@ -79,8 +79,8 @@ public final class StackOverflowClient implements Client {
 
     /* Retrieve feedback, if stored feedback exists */
     Entity feedback = getFeedback(url);
-    long upvote = (long) getUpvote(feedback);
-    long downvote = (long) getDownvote(feedback);
+    long upvote = getUpvote(feedback);
+    long downvote = getDownvote(feedback);
     return new Card(title, code, url, description, upvote, downvote);
   }
